Basics of DB
Two new packages were added to the server, aws-sdk and multer-s3. We access the s3 storage through multer, and the DB through sql.

In the index.js, whats been added was aws connections, DB connection status, multerS3 connection, image type collection, and DB query fixing.

In the uploadPage, I added a variable, fileType, to the Data object which saves whether the photo is a jpg, png, webp, etc. and sends that information to the backend. This will smooth the process when deleting the cropper section.