Moved the error messages/ made the map non existent + formatted doc
I was going to comment out the entire map but there was an issue where the error messages would not appear after i pressed submit, so I just made the map non existent and kept the "yes or no" prompt to get the error messages.

The error messages were both moved below the submit button .

I also formatted the document which might of changed some things.